backend/requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.4.0; python_version >= '3.8'
azure-core==1.30.2; python_version >= '3.8'
azure-eventhub==5.12.1; python_version >= '3.8'
azure-identity==1.17.1; python_version >= '3.8'
certifi==2024.7.4; python_version >= '3.6'
cffi==1.17.0; platform_python_implementation != 'PyPy'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
cryptography==43.0.0; python_version >= '3.7'
databricks-sdk==0.32.0; python_version >= '3.8'
exceptiongroup==1.2.2; python_version < '3.11'
fastapi==0.112.2; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
httptools==0.6.1
idna==3.8; python_version >= '3.6'
iniconfig==2.0.0; python_version >= '3.7'
msal==1.30.0; python_version >= '3.7'
msal-extensions==1.2.0; python_version >= '3.7'
numpy==1.23.5; python_version >= '3.10'
packaging==24.1; python_version >= '3.8'
pandas==1.5.3; python_version >= '3.8'
pluggy==1.5.0; python_version >= '3.8'
portalocker==2.10.1; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pydantic==2.8.2; python_version >= '3.8'
pydantic-core==2.20.1; python_version >= '3.8'
pyjwt[crypto]==2.9.0; python_version >= '3.8'
pytest==8.3.2; python_version >= '3.8'
python-dotenv==1.0.1
pyyaml==6.0.2
requests==2.32.3; python_version >= '3.8'
scipy==1.14.1; python_version >= '3.10'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
starlette==0.38.2; python_version >= '3.8'
tomli==2.0.1; python_version < '3.11'
typing-extensions==4.12.2; python_version >= '3.8'
urllib3==2.2.2; python_version >= '3.8'
uvicorn[standard]==0.30.6; python_version >= '3.8'
uvloop==0.20.0
watchfiles==0.23.0
websockets==13.0
